+++ src/config.rs
@@ -6,7 +6,10 @@ use serde::Deserialize;
 use crate::macros::ok_or_return;
 
 #[derive(Debug, Default, Deserialize)]
-pub struct Config {}
+pub struct Config {
+    #[serde(default)]
+    pub ephemeral: bool,
+}
 
 impl Config {
     pub fn load(path: &Path) -> Self {