+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Procedure when bumping the version number:
 
 ### Added
 - Config file
+- `ephemeral` config option
 
 ## v0.3.0 - 2022-08-22